fix: 🐛 moved independent entities to other modules
This commit will move remaining entities to others module
