+++ src/modules/visualization/other/entities/chart-dimension-item.entity.ts
