+++ src/modules/visualization/other/entities/chart-dimension.entity.ts
@@ -1,7 +1,7 @@
 import { Entity, JoinColumn, ManyToOne, OneToMany } from 'typeorm';
 
 import { ChartDimensionItem } from './chart-dimension-item.entity';
-import { Chart } from '../chart/entities/chart.entity';
+import { Chart } from '../../chart/entities/chart.entity';
 import { FavoriteDimension } from './favorite-dimension.entity';
 
 @Entity('chartdimension', { schema: 'public' })
