+++ src/modules/visualization/other/entities/favorite-dimension-item.entity.ts
