+++ src/modules/visualization/other/entities/favorite-dimension.entity.ts
