+++ src/modules/visualization/other/entities/map-view-dimension-item.entity.ts
