+++ src/modules/visualization/other/entities/map-view-dimension.entity.ts
