+++ src/modules/visualization/other/entities/report-table-dimension-item.entity.ts
